erp/requirements.txt
-i https://pypi.org/simple
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
django==2.1.3
djangorestframework==3.9.0
idna==2.7
psycopg2==2.7.6.1
python-decouple==3.1
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
